added creation of players using the factory, added taking care of GetBattleInfo actionrequest, some fixes in ourplayer

+++ Ex2/projFiles/unique/PlayerBattleInfo.h
@@ -7,21 +7,21 @@
 class PlayerBattleInfo : public BattleInfo{
 private:
     vector<vector<array<shared_ptr<matrixObject>, 3>>> tankGameBoard;
-    const size_t ClosestEnemyTankCol;
-    const size_t ClosestEnemyTankRow;
+    const int ClosestEnemyTankCol;
+    const int ClosestEnemyTankRow;
     int TurnsUntillNextUpdate;
     int TurnsFromLastUpdate;
     int numOfStartingTankBullets;
 
 public:
-    PlayerBattleInfo(size_t ClosestEnemyTankCol, size_t ClosestEnemyTankRow, int turnsUntillNextUpdate, 
+    PlayerBattleInfo(int ClosestEnemyTankCol, int ClosestEnemyTankRow, int turnsUntillNextUpdate, 
                                          vector<vector<array<shared_ptr<matrixObject>, 3>>> &tankGameBoard, int numOfStartingTankBullets)
         : tankGameBoard(tankGameBoard), ClosestEnemyTankCol(ClosestEnemyTankCol), ClosestEnemyTankRow(ClosestEnemyTankRow),
          TurnsUntillNextUpdate(turnsUntillNextUpdate), TurnsFromLastUpdate(-1), numOfStartingTankBullets(numOfStartingTankBullets) {
     }
     ~PlayerBattleInfo();
-    size_t getClosestEnemyTankCol() const { return ClosestEnemyTankCol; }
-    size_t getClosestEnemyTankRow() const { return ClosestEnemyTankRow; }
+    int getClosestEnemyTankCol() const { return ClosestEnemyTankCol; }
+    int getClosestEnemyTankRow() const { return ClosestEnemyTankRow; }
     int getTurnsUntillNextUpdate() const { return TurnsUntillNextUpdate; }
     void setTurnsUntillNextUpdate() { TurnsUntillNextUpdate--; }
     void setTurnsFromLastUpdate() { TurnsFromLastUpdate++; }
